refactor(request-executer): extract method that parses query params
